ObjectManager now controls objects
simplified code into object manager which runs all the object functions

+++ Rock.cpp
@@ -15,7 +15,7 @@ void Rock::Initialise(Vector2D position)
 	
 }
 
-void Rock::Update(double frametime)
+void Rock::Update(float frametime)
 {
 
 }
